Cleaned up
I added parameters for case length, height, width and wall thickness and
replaced many of the arbitraty numbers throughout the file with
references to these varaibles.

Also reduced the number of teeth in the arms from 13 to 9 since 4 of
them would never be used and it'll print a little faster.