Made more sturdy
Made the lip that holds the phone thicker and 3mm longer.
Added lips to the left and right racks to prevent the phone from falling
forward.
Added two arms to the bottom side of the lid to prevent the bottom side
of the case from bowing out under the weight of the phone.